Use stryker api (#110)
* refactor(api) Remove api from main stryker repo

* Add peerDependency to stryker-api
* Add devDepencency to stryker-api, make sure it is installed during dev
* Renamed all imports to 'stryker-api/...'

* fix(test-selector-api) Implement new test selector api

The test selector API was pretty complicated, test selectors were also
responsible for the file IO. Now moved that responsibility to the
TestRunnerOrchestrator
* Test selector now only is responsible for delivering file content for
  selecting tests
* Altered TestRunnerOrchestrator logic to accomondate for it

* refactor(stryker) Replace piramid with then() chain

The stryker runMutationTest() method was written like a piramid. This made
it so that we had to catch every error and reject the promise by hand. By
chaining promises, we ensure that errors thrown result in the rejection.

* feat(reporter-wrap-up) Implement 'wrapUp'

Implement a new feature on reporters. The wrapUp function gives a reporter
the change to finish up any (sync or async) work.