fix(StrykerTempFolder): Use local tmp folder (#121)
* fix(StrykerTempFolder): Use local tmp folder

When running mocha tests, we need to be able to resolve the local node_modules. The way this works in npm is that it searches for a node_modules in de current working directory. If not found, check the dir above. Etc. By creating the tmp-folder's locally, node_,modules will be resolved as expected.

* fix(stryker-temp): Clean temp folder

* fix(delete-dir): Fix for deleting a dir recrusive