chore: Add yarn.lock file
yarn is an alternate npm client that can install packages
much faster than the npm client in most benchmarks. To achieve that
speed, it needs to keep a record of package dependencies so that
results are consistent between machines and so that caching can be maximised.
The dependency tree data goes in a yarn.lock file.

Having the yarn.lock file will make it easier for contributors to
install required dependencies.