feat(basic-runner): Implementation of a basic Jest-runner (#4)
* chore: remove some template files from the plugin-seed
* chore: update project files for newer Stryker version
* chore(doc): Show correct Travis badge
* test(all): Add sample project for running integration tests
* fix(build) Explicitly list jest-runtime as dep
* fix(TestRunner): Add testPathDirs option
Add `testPathDirs` option to the jest test runner options in order to make it work on the windows platform. I'm not jet sure that `process.cwd()` is the correct value, but it seems to work for now.
* chore(build) Do not attempt to run on Node < 6.0.0
* fix(runner) Also specify `roots` option
This in preparation of future Jest versions (> 18.1) that
will have `testPathDirs` renamed to `roots`.
See https://github.com/facebook/jest/pull/2776.