test(own-dogfood): Add stryker-api to stryker files
When running stryker on stryker, we didn't include the stryker-api module in the test files. This meant that every sandbox shared the stryker-api files with the `IsolatedtestRunnerAdapterWorker` child process. Since we now have a `MochaConfigWriter`, that one is always added to the `ConfigWriterFactory`. It poluted the tests in the sandbox, resulting in an initial test failure.