fix(webpack-transpiler): Correctly use in-memory files (#601)
Correctly use in-memory files. Instead of caching them ourselves, extend webpack's `CachedInputFileSystem` and rely on that for caching files. Still use memory-FS to load the in-memory files.

One downside is that I had to set the typescript target to "es2015", which means we have to drop support for node-4 on the stryker-webpack-transpiler package.

Fixes #597