fix(Webpack transpiler): Loader exception logging (#700)
The webpack transpiler, when trying to load a config file, incorrectly
swallowed exceptions caused by errors in the config file.

Instead of logging the proper error you would get a misleading file not
found error.

To help the user, we perform an explicit check on the existence of
the configuration file and leave the rest up for the standard behavior
of node. It turns out that this is good enough for people to debug the
problem on their own.