fix(any-promise): Don't register a promise implementation (#1180)
Remove dependency on 'mz' which in turn removes the depencency
on 'any-promise'. This will allow test processes to roll their own
Promise implementation
using ['any-promise'](https://www.npmjs.com/package/any-promise)

Add '@stryker-mutator/util' as a dependency. In there we can maintain library
logic that we want to share between plugins and core packages.
The promisified fs and child_process implementations moved to this package.
The idea of promisified API's is that they can gradually
be removed in the future (node 8 and 10).