feat(Conditional expression mutator): Mutate conditional operators (#1253)
The `ConditionalExpressionMutator` now also mutates conditional expressions in the wild. Not only in `for`, `while` and `if` statements.

For example: `const foo = bar < baz` will be mutated to `const foo = bar <= baz`. 

Loops are still not mutated to obvious infinite loops (for example `while (i < 10) { i++ }` **isn't** mutated to `while (true) { i++ }`)