feat(dependency injection): Add dependency injection for plugins (#1313)
This PR adds dependency injection to Stryker, in preparation of removing the side effects from the way we now load plugins. Instead of registering yourself to the correct factory (for example, by calling `ReporterFactory.instance().register`), it relies on plugins exporting a property called `strykerPlugins`. 

Just to be clear: this PR does fix not the issue reported in #667. It only adds the dependency injection mechanism to Stryker. We still need to do PR's for each package (can go relatively fast)

The content of this PR:

1. Add a dependency injection framework called `typed-inject` to the packages folder; A type safe dependency injection framework
1. Add a package `stryker-api/di` which contains dependency injection related interfaces to help plugins know what to (type-safely) inject
1. Add `typed-inject` as a dependency to the core Stryker package and use it. Reporter plugins are loaded via this new mechanism.
1. Alter the `PluginLoader` in order to enable dependency injection in all plugins. It is still backward compatible with the old way of loading plugins.
1. Add a package to contain generic test helpers called `@stryker-mutator/test-helpers`. This package contains a `testInjector` to help inject common stuff into injectables for testability purposes.
1. It updated the `BroadcastReporter` and all build-in reporters to now use the new dependency injection mechanism.