feat(HtmlReporter): include the html reporter in the core package and add it to the default reporters (#2036)
Adds the html reporter to the list of default reporters.
The `@stryker-mutator/html-reporter` package is now deprecated and only a stub. There is no need to install this package anymore since the html reporter is bundled with the `@stryker-mutator/core` package.

Fixes #1919 

BREAKING CHANGE: the `html` reporter is now enabled by default. If you don't want to use it, be sure to override the `reporters` configuration option.

```json
{
  "reporters": ["progress", "clear-text"]
}
```