chore: publish new version via github workflow (#2075)
Publish new releases via a GitHub workflow. This will allow us to remove the last bit of Travis ci.

The new way to release Stryker: Run `npx lerna version` locally on master. This allows you to override the versioning (although the default is based on conventional commit). The publish workflow should take care of the rest.