feat(core): support build command
Add support for `buildCommand` configuration option. This allows users to configure a command which is run inside the sandbox just before actual mutation testing starts, but after the files are mutated.

This can be used to run transpilation or bundling for setups where the test runner doesn't already take care of that.

For example:

```json
{
  "buildCommand": "npm run build"
}
```

Chore:
* Add dependency on `execa` and `npm-run-path` to help with the running of a command
* Enable 2 more e2e tests: `vue-javascript` nad `webpack-zero-conf-karma`