feat(test-framework): remove `TestFramework` API
Remove the test framework API and their implementations: `@stryker-mutator/jasmine-framework` and `@stryker-mutator/mocha-framework`. Test runners are now expected to cary their own support for different test frameworks they might support.

A `TestFramework` provided decoupling between the framework and the test runner itself. It was used to provide test filtering and `perTest` coverage analysis. Historically, this decoupling only came in use for `karma`, since that is the only test runner that supports multiple test frameworks. Support for filtering `mocha` and `jasmine` tests is now build into the `@stryker-mutator/karma-runner` itself, as well as the other supported test runners.

Therefore this API is no longer used, which makes it easier to configure Stryker and easier to maintain.