feat(command): Support command test runner with mutation switching
Add support for the command test runner with mutation switching.

_NOTE: Support for the command test runner is best effort!_. It might not work for your test runner of choice.

Switching of mutants is done by providing the __STRYKER_ACTIVE_MUTANT__ environment variable when executing the command. The `@stryker-mutator/instrumenter` now adds a check for this environment variable in each source file. This only works if the command you configure allows for this environment variable to flow through to the test environment.

In other words: this should work for node test runners like [node-tap](https://node-tap.org/) and mocha but not for test runners that spin up their own test environment, like jest or karma.

Note that for mocha, karma and jest, you're better of using their respective test runner plugins.