feat(core): add ability to override file headers
Add the ability to override the file headers inside a sandbox. This allows users to opt-out of the default header, as well as add custom headers.

You can configure it like this:

```json
{
  "sandboxFileHeaders": {
  "**/*.ts": "// Hello world\n"
  }
}
```

The key here is a [glob expression](https://globster.xyz/), where the value points to the header to be used for matching files.