feat(test runner api): remove `sandboxFileNames` injectable values (#2369)
Remove `sandboxFileNames` as injectable value for test runner plugins. Test runner plugins are expected to do their own file discovery in the current working directory. The list of sandbox files that Stryker internally copied over might differ from the list of files in the sandbox, because the user might have configured a `buildCommand` that compiled the files to their final representation.

* Remove `sandboxFileNames` from dependency injection context
* Rename `OptionsContext` to `PluginContext` and remove the `PluginContexts` mapped type (no longer needed)
* Move generic require cache clean functionality to `@stryker-mutator/util` 
    * Use this in the mocha-runner 
    * Use this in the jasmine-runner.

Fixed #2351