feat(reporter-api): support mutation switching
Update the `Reporter` api in order to support mutation switching.

* Implement `MutantResult` as a union type
* Add `killedBy` to `KilledMutantResults`.
* Add `errorMessage` to `InvalidMutantResult`.
* Add the `testFilter` to `InvalidMutantResult`.
These names are in line with the [stryker handbook](https://github.com/stryker-mutator/stryker-handbook/blob/master/mutant-states-and-metrics.md).

Renamed `sourceFilePath` to `fileName` to make it more inline with the naming in the rest of the api.

BREAKING CHANGE: The `onMutantTested` and `onAllMutantsTested` methods on the `Reporter` api have changed