feat(test-runner): Add `--maxTestRunnerReuse` support
Add support for `--maxTestRunnerReuse` compiler flag (and equivalent in stryker.conf.json). With it, you can specify the maximum number that a test runner can be reused before the process is killed and a new one is created. The default is `0`, which means no maximum, just keep reusing.

There are 2 use cases for this setting.

1. You have a memory leak in your tests that cannot be fixed (for example, a plugin in jest). You run with `--maxTestRunnerReuse 20`.
2. Your environment is dirty after running one time and you cannot do anything about it. You run with `--maxTestRunnerReuse 1`.