feat(instrumenter): add excludedMutations support (#2513)
Add support to configure `mutator.excludedMutations`. This was already supported in Stryker@<3, but not yet in Stryker4 beta.

Instead of not generating the mutants, it generates them without placing them in the code. They get reported as `ignored` in your mutation testing report.

![image](https://user-images.githubusercontent.com/1828233/94722119-2142dc00-0357-11eb-839c-a4aef430dcea.png)

Also, add error message, ignore reason, and "killed by" test name in the "show more" modal dialog.

![image](https://user-images.githubusercontent.com/1828233/94722368-8565a000-0357-11eb-9a2d-05be1ac0d161.png)