feat(debugging): allow passing node args to the test runner
Add the `--testRunnerNodeArgs` option. With this option, you can pass arguments to the node process that hosts the test runner.

For example, running stryker with `--timeoutMS 9999999 --concurrency 1 --testRunnerNodeArgs --inspect-brk` will allow you to debug the tests running in the dryRun or the mutation test run.

Another use case is using `--testRunnerNodeArgs --cpu-prof`, this allows you to analyse the performance of the child process..