refactor(pool): implement resource-like pattern
Implement resource like pattern. Schedule work to be done on the pool with `pool.schedule(input$, task)`.

The pool itself is responsible to make sure only one task is executed on one worker at any one time. The pool is also responsible to recycle the worker after it is done with a task.