feat(mutation range): allow specifying a mutation range (#2751)
Add possibility to specify a mutation range with `mutate`. For example:

```json
{
  "mutate": ["src/app.js:1-20"]
}
```

This mutates the first 20 lines in app.js only.

The syntax is: `:startLine[:startColumn]-endLine[:endColumn]`, where columns are optional. 

It is not possible to use a mutation range on a glob expression. For example: using `"src/*.js:1-20:` results in an error message.