fix(typescript-checker): support TS 4.4 (#3178)
TypeScript 4.4 also watches package.json files in node_modules for changes. I guess this is done to be able to reload the types when a node_module gets updated.

However, the typescript-checker wasn't expecting those `watchFile` calls to come in. The checker was defensive programmed, meaning an error is thrown when something happens that it doesn't expect.

This ended up being a simple fix. Simply ignoring `watchFile` of files that are not loaded seems to do the trick.