feat(ignore static): allow to ignore static mutants (#3284)
Add support for `--ignoreStatic`. When `ignoreStatic` is enabled, static mutants are ignored (defaults to `false`).

Static mutants can be expensive, depending on your use case. For example, 6% of mutants of `@stryker-mutator/core` are static, yet they take up > 50% of time executing tests. So it might make sense to ignore them.