feat(reload test environment): implement test environment reload  (#3369)
Add the ability for Stryker to restart a test runner in order to reload its test environment. This is used to test static mutants where the test runner is incapable of reloading the test environment by itself. This is a requirement to support es modules (ESM) since it is impossible to reload an ESM using something like `delete require.cache['...']`. Instead, Stryker will restart the entire test runner process for those mutants.

The way for the test runner to communicate to Stryker whether it is capable of reloading the environment is by implementing the new `capabilities()` method.

```ts
capabilities() {
  return { reloadEnvironment: true }
}
```

If you do this, you can expect the new `reloadEnvironment` boolean in your `MutantRunOptions` to be true when you need to reload the test environment after activating the mutant.

BREAKING CHANGE: Test runner plugins must provide `TestRunnerCapabilities` by implementing the `capabilities` method.