fix(core): report failed check results only once
Share the replay of failed check results accross all subscribers using the `shareReplay()` rxjs operator.

Reporting of a failed checks is done using a side effect (with the `tap()` operator). When later we subscribe to it twice, which is a result of the `partition()` operator, the reporting was done twice.