chore(lock): lock dependencies (#3606)
Lock all dependencies using package-lock files. We've gone back and forth on this. The last time we removed the package-lock files we did so because we were unhappy with dependabot's version bumps. We now use renovate bot and they [recommend using a lock file](https://docs.renovatebot.com/dependency-pinning/#so-whats-best).

I hate a failing publish pipeline for the simple reason of babel pushing an update to their types 🙈

- chore(babel-types): support latest babel types definition
- chore(lock): lock dependencies