feat(temp-dir): allow multiple instances to share the same temp-dir (#5120)
* feat(temp-dir): allow multiple instances to share the same temp-dir

Allow multiple runs of Stryker to share the same temp dir. This means that Stryker will _no longer remove the entire `.stryker-tmp` directory when it ran successfully (since other `sandbox-xxx` directories could be owned by other processes).

Note: this is implemented in preparation for the implementation of the mutation server protocol (#5086)

* fix test on windows