fix(child-process): fix OOM detection on apple silicon
Capture 8kB of process output for detecting OOM instead of 2kB.

The out of memory detection works by scanning the process output in case of a crash. Max 2kB of output is scanned. However, when the process crashes on Apple Silicon, the error output contains a stacktrace that pushes the "JavaScript heap out of memory" message over the 2kB.